requirements.txt
flask==2.0.1
google-generativeai>=0.3.0
python-dotenv>=0.19.0
flask-caching>=1.10.0
waitress>=2.1.0